Add server-side validation

+++ server/server.lua
@@ -33,19 +33,30 @@ RSGCore.Commands.Add('resetsaloon', locale('sv_lang_7'), { { name = 'saloonid', 
 
 end, 'admin')
 
+
+---------------------------------------------
+-- functions
+---------------------------------------------
+function isPlayerSaloonOwner(src, saloonid)
+    local Player = RSGCore.Functions.GetPlayer(src)
+    if not Player then return false end
+    local result = MySQL.query.await('SELECT * FROM rex_saloon WHERE owner = ? AND saloonid = ?', { Player.PlayerData.citizenid, saloonid})
+    return result[1] ~= nil
+end
+function countOwnedSaloons(src)
+    local Player = RSGCore.Functions.GetPlayer(src)
+    if not Player then return 0 end
+    local citizenid = Player.PlayerData.citizenid
+    local result = MySQL.prepare.await("SELECT COUNT(*) as count FROM rex_saloon WHERE owner = ?", { citizenid })
+    return result
+end
+
 ---------------------------------------------
 -- count owned saloons
 ---------------------------------------------
 RSGCore.Functions.CreateCallback('rex-saloon:server:countowned', function(source, cb)
     local src = source
-    local Player = RSGCore.Functions.GetPlayer(src)
-    local citizenid = Player.PlayerData.citizenid
-    local result = MySQL.prepare.await("SELECT COUNT(*) as count FROM rex_saloon WHERE owner = ?", { citizenid })
-    if result then
-        cb(result)
-    else
-        cb(nil)
-    end
+    cb(countOwnedSaloons(src))
 end)
 
 ---------------------------------------------
@@ -80,6 +91,18 @@ end)
 RegisterNetEvent('rex-saloon:server:newstockitem', function(saloonid, item, amount, price)
     local src = source
     local Player = RSGCore.Functions.GetPlayer(src)
+
+    if not isPlayerSaloonOwner(src, saloonid) then
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_not_saloon_owner'), type = 'error', duration = 7000 })
+        return
+    end
+
+    if not Player.Functions.RemoveItem(item, amount) then
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_no_item'), type = 'error', duration = 7000 })
+        return
+    end
+    
+    TriggerClientEvent('rsg-inventory:client:ItemBox', src, RSGCore.Shared.Items[item], 'remove', amount)
     local itemcount = MySQL.prepare.await("SELECT COUNT(*) as count FROM rex_saloon_stock WHERE saloonid = ? AND item = ?", { saloonid, item })
     if itemcount == 0 then
         MySQL.Async.execute('INSERT INTO rex_saloon_stock (saloonid, item, stock, price) VALUES (@saloonid, @item, @stock, @price)',
@@ -89,15 +112,11 @@ RegisterNetEvent('rex-saloon:server:newstockitem', function(saloonid, item, amou
             ['@stock'] = amount,
             ['@price'] = price
         })
-        Player.Functions.RemoveItem(item, amount)
-        TriggerClientEvent('rsg-inventory:client:ItemBox', src, RSGCore.Shared.Items[item], 'remove', amount)
     else
         MySQL.query('SELECT * FROM rex_saloon_stock WHERE saloonid = ? AND item = ?', { saloonid, item }, function(data)
             local stockupdate = (amount + data[1].stock)
             MySQL.update('UPDATE rex_saloon_stock SET stock = ? WHERE saloonid = ? AND item = ?',{stockupdate, saloonid, item})
             MySQL.update('UPDATE rex_saloon_stock SET price = ? WHERE saloonid = ? AND item = ?',{price, saloonid, item})
-            Player.Functions.RemoveItem(item, amount)
-            TriggerClientEvent('rsg-inventory:client:ItemBox', src, RSGCore.Shared.Items[item], 'remove', amount)
         end)
     end
 end)
@@ -105,29 +124,60 @@ end)
 ---------------------------------------------
 -- buy item amount / add money to account
 ---------------------------------------------
-RegisterNetEvent('rex-saloon:server:buyitem', function(amount, item, newstock, price, label, saloonid)
+RegisterNetEvent('rex-saloon:server:buyitem', function(amount, item, saloonid)
     local src = source
+
+    local Player = RSGCore.Functions.GetPlayer(src)
+    if not Player then return end
 
     if amount <= 0 then
         TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_invalid_amount'), type = 'error', duration = 7000 })
         return
     end
 
-    local Player = RSGCore.Functions.GetPlayer(src)
+    -- Get stock data / verify existence
+    local result = MySQL.query.await('SELECT id, stock, price FROM rex_saloon_stock WHERE saloonid = ? AND item = ?', {saloonid, item})
+    if not result[1] then
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_stock_not_found'), type = 'error', duration = 7000 })
+        return
+    end
+    local stockId = result[1].id
+    local stockPrice = result[1].price
+    local stockAmount = result[1].stock
+
+    -- Verify stock amount
+    if stockAmount < amount then
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('Not enough in stock'), type = 'error', duration = 7000 })
+        return
+    end
+
+    -- Verify player money
     local money = Player.PlayerData.money[Config.Money]
-    local totalcost = (price * amount)
-    if money >= totalcost then
-        MySQL.update('UPDATE rex_saloon_stock SET stock = ? WHERE saloonid = ? AND item = ?', {newstock, saloonid, item})
+    local totalcost = (stockPrice * amount)
+    if money < totalcost then
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_lang_1')..Config.Money, type = 'error', duration = 7000 })
+        return
+    end
+
+    -- Sell item to player and update saloon data
+    if Player.Functions.AddItem(item, amount) then
+        TriggerClientEvent('rsg-inventory:client:ItemBox', src, RSGCore.Shared.Items[item], 'add', amount)
         Player.Functions.RemoveMoney(Config.Money, totalcost)
-        if Player.Functions.AddItem(item, amount) then
-            TriggerClientEvent('rsg-inventory:client:ItemBox', src, RSGCore.Shared.Items[item], 'add', amount)
-        end
+        
+        -- Update saloon money
         MySQL.query('SELECT * FROM rex_saloon WHERE saloonid = ?', { saloonid }, function(data2)
             local moneyupdate = (data2[1].money + totalcost)
             MySQL.update('UPDATE rex_saloon SET money = ? WHERE saloonid = ?',{moneyupdate, saloonid})
         end)
-    else
-        TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_lang_1')..Config.Money, type = 'error', duration = 7000 })
+
+        -- Update saloon item
+        local newStockAmount = stockAmount - amount
+        if newStockAmount > 0 then
+            MySQL.update('UPDATE rex_saloon_stock SET stock = ? WHERE id = ?', { newStockAmount, stockId })
+        else
+            print('delete item from stock, id of stock', stockId)
+            MySQL.Async.execute('DELETE FROM rex_saloon_stock WHERE id = ?', { stockId })
+        end
     end
 end)
 
@@ -137,9 +187,16 @@ end)
 RegisterNetEvent('rex-saloon:server:removestockitem', function(data)
     local src = source
     local Player = RSGCore.Functions.GetPlayer(src)
+
+    if not isPlayerSaloonOwner(src, data.saloonid) then
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_not_saloon_owner'), type = 'error', duration = 7000 })
+        return
+    end
+
     MySQL.query('SELECT * FROM rex_saloon_stock WHERE saloonid = ? AND item = ?', { data.saloonid, data.item }, function(result)
-        Player.Functions.AddItem(result[1].item, result[1].stock)
-        TriggerClientEvent('rsg-inventory:client:ItemBox', src, RSGCore.Shared.Items[result[1].item], 'add', result[1].stock)
+        if Player.Functions.AddItem(result[1].item, result[1].stock) then
+            TriggerClientEvent('rsg-inventory:client:ItemBox', src, RSGCore.Shared.Items[result[1].item], 'add', result[1].stock)
+        end
         MySQL.Async.execute('DELETE FROM rex_saloon_stock WHERE id = ?', { result[1].id })
     end)
 end)
@@ -163,6 +220,12 @@ end)
 RegisterNetEvent('rex-saloon:server:withdrawfunds', function(amount, saloonid)
     local src = source
     local Player = RSGCore.Functions.GetPlayer(src)
+
+    if not isPlayerSaloonOwner(src, saloonid) then
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_not_saloon_owner'), type = 'error', duration = 7000 })
+        return
+    end
+
     MySQL.query('SELECT * FROM rex_saloon WHERE saloonid = ?', {saloonid} , function(result)
         if result[1] ~= nil then
             if result[1].money >= amount then
@@ -182,7 +245,23 @@ RegisterNetEvent('rex-saloon:server:rentsaloon', function(saloonid)
     local Player = RSGCore.Functions.GetPlayer(src)
     local money = Player.PlayerData.money[Config.Money]
     local citizenid = Player.PlayerData.citizenid
+
+    local saloonData = MySQL.query.await('SELECT * FROM rex_saloon WHERE saloonid = ?', { saloonid })[1]
+    if not saloonData then
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('invalid_saloon'), type = 'error', duration = 7000 })
+        return 
+    end
+    if saloonData.owner ~= 'vacant' then
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('saloon_already_rented'), type = 'error', duration = 7000 })
+        return
+    end
+
     if money > Config.RentStartup then
+        if countOwnedSaloons(src) >= Config.MaxSaloons then
+            TriggerClientEvent('ox_lib:notify', src, {title = locale('cl_lang_48'), description = locale('cl_lang_49'), type = 'error', duration = 7000 })
+            return
+        end
+
         Player.Functions.RemoveMoney(Config.Money, Config.RentStartup)
         Player.Functions.SetJob(saloonid, 2)
         if Config.LicenseRequired then
@@ -204,16 +283,25 @@ end)
 RegisterNetEvent('rex-saloon:server:addrentmoney', function(rentmoney, saloonid)
     local src = source
     local Player = RSGCore.Functions.GetPlayer(src)
+
+    if not isPlayerSaloonOwner(src, saloonid) then
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_not_saloon_owner'), type = 'error', duration = 7000 })
+        return
+    end
+
     MySQL.query('SELECT * FROM rex_saloon WHERE saloonid = ?', { saloonid }, function(result)
         local currentrent = result[1].rent
         local rentupdate = (currentrent + rentmoney)
         if rentupdate >= Config.MaxRent then
             TriggerClientEvent('ox_lib:notify', src, {title = 'Can\'t add that much rent!', type = 'error', duration = 7000 })
         else
-            Player.Functions.RemoveMoney(Config.Money, rentmoney)
-            MySQL.update('UPDATE rex_saloon SET rent = ? WHERE saloonid = ?',{ rentupdate, saloonid })
-            MySQL.update('UPDATE rex_saloon SET status = ? WHERE saloonid = ?', {'open', saloonid})
-            TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_lang_4'), type = 'success', duration = 7000 })
+            if Player.Functions.RemoveMoney(Config.Money, rentmoney) then
+                MySQL.update('UPDATE rex_saloon SET rent = ? WHERE saloonid = ?',{ rentupdate, saloonid })
+                MySQL.update('UPDATE rex_saloon SET status = ? WHERE saloonid = ?', {'open', saloonid})
+                TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_lang_4'), type = 'success', duration = 7000 })
+            else
+                TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_not_enough_money'), type = 'error', duration = 7000 })
+            end
         end
     end)
 end)
@@ -221,22 +309,24 @@ end)
 ---------------------------------------------
 -- check player has the ingredients
 ---------------------------------------------
+local function hasIngredients(src, ingredients)
+    local icheck = 0
+    for k, v in pairs(ingredients) do
+        if exports['rsg-inventory']:GetItemCount(src, v.item) < v.amount then
+            return false
+        end
+    end
+    return true
+end
+
 RSGCore.Functions.CreateCallback('rsg-saloon:server:checkingredients', function(source, cb, ingredients)
     local src = source
-    local hasItems = false
-    local icheck = 0
     local Player = RSGCore.Functions.GetPlayer(src)
-    for k, v in pairs(ingredients) do
-        if exports['rsg-inventory']:GetItemCount(src, v.item) >= v.amount then
-            icheck = icheck + 1
-            if icheck == #ingredients then
-                cb(true)
-            end
-        else
-            TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_lang_5'), type = 'error', duration = 7000 })
-            cb(false)
-            return
-        end
+    if hasIngredients(src, ingredients) then
+        cb(true)
+    else
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_lang_5'), type = 'error', duration = 7000 })
+        cb(false)
     end
 end)
 
@@ -248,6 +338,12 @@ RegisterNetEvent('rex-saloon:server:finishcrafting', function(data)
     local Player = RSGCore.Functions.GetPlayer(src)
     local receive = data.receive
     local giveamount = data.giveamount
+
+    if not hasIngredients(src, data.ingredients) then
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_lang_5'), type = 'error', duration = 7000 })
+        return
+    end
+
     for k, v in pairs(data.ingredients) do
         Player.Functions.RemoveItem(v.item, v.amount)
         TriggerClientEvent('rsg-inventory:client:ItemBox', src, RSGCore.Shared.Items[v.item], 'remove', v.amount)
@@ -275,6 +371,12 @@ RegisterServerEvent('rex-saloon:server:storagebrewing', function(saloonid)
     local src = source
     local Player = RSGCore.Functions.GetPlayer(src)
     if not Player then return end
+
+    if not isPlayerSaloonOwner(src, saloonid) then
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_not_saloon_owner'), type = 'error', duration = 7000 })
+        return
+    end
+
     local data = { label = 'Saloon Brewery', maxweight = Config.BrewingMaxWeight, slots = Config.BrewingMaxSlots }
     local stashName = 'brewery_'.. saloonid
     exports['rsg-inventory']:OpenInventory(src, stashName, data)
@@ -287,6 +389,12 @@ RegisterServerEvent('rex-saloon:server:storagestock', function(saloonid)
     local src = source
     local Player = RSGCore.Functions.GetPlayer(src)
     if not Player then return end
+
+    if not isPlayerSaloonOwner(src, saloonid) then
+        TriggerClientEvent('ox_lib:notify', src, {title = locale('sv_not_saloon_owner'), type = 'error', duration = 7000 })
+        return
+    end
+
     local data = { label = 'Saloon Stock', maxweight = Config.StockMaxWeight, slots = Config.StockMaxSlots }
     local stashName = 'stock_'.. saloonid
     exports['rsg-inventory']:OpenInventory(src, stashName, data)
